add mutex for local chain
